Move & update publish code to `planning.publish`
